+++ lib/components/i18n/index.d.ts
@@ -1,0 +1,3 @@
+export * from './es';
+export * from './en';
+//# sourceMappingURL=index.d.ts.map